Add configure sheet to screen saver
I'm not sure why all of this works.

I looked at how https://github.com/burnsra/KrogerSecurityScreenSaver
does it, and they used a plain object instead of a window controller.
It's essentially a singleton because it keeps a sharedInstance that is
used to access everything. It retains an NSWindow. On init, it loads
the custom nib from the bundle.

To access the config window, you get the sharedInstance and then call
it's window.

To close the window, it calls endSheet on the sheetParent of configure
window.

The other change was getting rid of the detached window bug that I was
having. The bug was that when you opened the preferences window, it
would place it self weirdly detached from the main window instead of
showing itself as a sheet. When you re-opened the window, it was then
attached. The fix was disabling the "Visible at launch" setting for the
window.

My guess is this fixed the problem because when the window was trying
to display at launch, it wasn't attached to a parent window yet. So
when I "opened" it for the first time, it was already open and just
came to the front, still detached.

Now, it makes sure it only opens when the user wants it and therefore
is attached to something that exists.
