+++ ColorClockSaver/MainView.swift
@@ -5,7 +5,6 @@ class MainView: ScreenSaverView {
   let wrapperView = NSStackView()
   let timeView = TimeView()
   let colorCodeView = ColorCodeView()
-  lazy var configureSheetController = ConfigureSheetController()
 
   override func viewDidMoveToWindow() {
     wantsLayer = true
@@ -64,6 +63,7 @@ class MainView: ScreenSaverView {
   }
 
   override func configureSheet() -> NSWindow? {
-    return configureSheetController.window
+    let configureSheet = ConfigureSheet.sharedInstance
+    return configureSheet.window
   }
 }
